requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PySide6
requests
python-dotenv
pydantic
uvicorn
fastapi
pinecone
sentence-transformers
ollama
fastapi-mcp
duckduckgo-search
plyer
winsdk
openwa
torch
msal 